+++ hidden/Common/Constant.swift
@@ -1,6 +1,6 @@
 //
 //  Constant.swift
-//  vanillaClone
+//  Constant.swift
 //
 //  Created by Thanh Nguyen on 1/30/19.
 //  Changed by Vitalii Tereshchuk / xVoLAnD, 2026